Updates display director name.
The display name for directors is in the format Lastname, Firstname.
For movies with more than one director, it takes the comma out of the
middle of the list, but both names are still shown in correct order.

The list of directors is also now sorted alphabetically by last name.